Fix: lib: do not check the frozen state in bt_X_set_is_frozen()
Some freezing functions were called bt_X_freeze() and were transformed
into freezing/thawing functions bt_X_set_is_frozen(), accepting the
frozen state as a parameter. The optimization checks to avoid refreezing
when the object is already frozen were left, but we can thaw now, so we
must not check this.

It could be optimized again, but those specific freezing functions are
only enabled in developer mode, so it's not a complete catastrophe.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>