Make the bt2 Python bindings build
This patch is just a little bit more than the bare minimum to make the
Python bindings build (don't try to use them yet, though).  It consists
of sync'ing the .i SWIG interface definition files with the
corresponding headers.  It's possible that the interfaces are
incompleted, in which case they'll be adjusted in the a later pass.  For
example, output parameters are currently not named in a way that match
the type maps.  But this patch contains the bulk of the changes.

For simplicity, I have included at the same time some other changes to
the .i files, such as adjusting the typemap definitions (in
native_bt.i).  The helper functions (such as at the bottom of
native_bt_componentclass.i) are also changed to their "final" form.  While
these are not immediatly used, they needed to be changed for the
bindings to build anyway.

Signed-off-by: Simon Marchi <simon.marchi@efficios.com>