Add env wrapper for bt2 Python bindings tests
The test_python_bt2 script sets up a special environment for running the
tests.  It can be difficult to reproduce this environment exactly when
working on a single test.

This patch introduces test_prython_bt2_env, which can be used to run a
command in the same environment.  This command can be a single test run,
or even a sub-shell.

Once the bindings and their tests work better, we should make
test_python_bt2 use test_python_bt2_env, to avoid the duplication, but
for now I prefer to leave it as-is.

Signed-off-by: Simon Marchi <simon.marchi@efficios.com>