Make the bt2 and bt2.ctfwriter packages importable
The goal of this patch is to make it possible to do "import bt2",
nothing more.  It renames and removes a few things.  There are still a
lot of references to things that don't exist anymore, but they are not
exercised by just importing the package, so I have not touched them.

* field_path.py: Remove references to packet and event header, concepts
  that don't exist in trace-ir anymore.

* field_types.py: s/FIELD_TYPE_ID/FIELD_CLASS_TYPE/

* fields.py: Likewise.

* notification.py: s/NOTIFICATION_TYPE_/MESSAGE_TYPE_/, and _BEGIN ->
  _BEGINNING.  NOTIFICATION_TYPE_INACTIVITY becomes
  MESSAGE_TYPE_MESSAGE_ITERATOR_INACTIVITY.

* VALUE_TYPE_FLOAT becomes VALUE_TYPE_REAL.

Signed-off-by: Simon Marchi <simon.marchi@efficios.com>