tests/cli: Add CLI params parsing test
This patch adds a test to verify the parsing of the parameters passed to
components from the Babeltrace command line.  The
bt_plugin_test_params.py file contains a sink component class that
prints whatever parameters it receives on initialization.  In the
test_params test, we pass some parameters and validate that the
component received what we expected.