tests/cli: Remove false positive in test_convert_args.in
Test 'path leftover + user filter named `muxer`' is missing an argument
to test_bt_convert_run_args, but is still reported as a pass.

The missing argument is the second one, the arguments to "babeltrace
convert".  What is supposed to be the third argument (expected output)
becomes the second argument (the arguments passed to "babeltrace
convert") and the actual third argument (expected output) is empty.  The
call to "babeltrace convert" fails because it is passed arguments it
doesn't know about, and nothing is output on its stdout.  We then
compare the process output (empty string) to the actual expected output
(also empty string) and the test passes.

I added a check for the exit code of babeltrace as a safety net that
would have caught this.  I just disabled the test for now, because I am
not sure what the test is trying to test, and what the convert args
should really be.