Update Python object model
This patch updates the Babeltrace Python object model (what all
Babeltrace Python objects all inherit from) to match the current status
of the API.

What have both Shared and Unique objects.

A Shared object type is defined by inheriting from _SharedObject.  It
wraps and owns a reference to a Babeltrace object that is reference
counted.  The reference count is dropped when the Python object gets
deleted.  The _SharedObject base class defines two class methods,
_get_ref and _put_ref, which must be implemented by subclasses.  These
implementations must respectively get and put a reference of the
specialized pointer they wrap.

A Unique object type is defined by inheriting from _UniqueObject.  It is
not refcounted.  Rather, it is wholly owned by some object (the owner,
or parent) that is itself refcounted.  A Unique object, in the
Babeltrace API, is destroyed when its parent is destroyed.  There is
therefore a risk that someone would keep a (Python) reference to a
unique object that got destroyed, because its parent got destroyed.  To
avoid this, we make it so that when one acquires a (Python) reference on
a (Python) Unique object, it obtains a reference on the parent Shared
object.  When the Unique Python object is destroyed, the reference on
the Shared parent is dropped.

The concepts of Freezable and Private objects don't exist anymore, so
they are removed.  I have also removed the _PrivateConnection class,
since it won't be needed in the end and inherited from _PrivateObject.

Change-Id: I9ab743dd7a28407913945648c1d74c467550df14
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>