cli: remove `convert` command's --path and --url options
It was decided that these two options can be removed.  They are
shortcuts for things that can be set using --params.

The --path option, in particular, is not useful as it used to be: the
de-facto standard parameter that sources are going to accept for their
input data is no longer `path` (it is now `inputs`).

Change-Id: I023e9cdaf319d7a9d2951315ce595ce099648391
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>