cli: set log levels earlier
The problem this patch is aiming to solve is that the effective log
level is not set when executing bt_config_convert_from_args.  For
example, bt_python_bindings_bt2_log_level is still NONE, only to be set
later.  So if Python plugin fails to be imported during the auto source
discovery phase, nothing will be logged, making debugging quite
difficult.  Any argument passed (top level --log-level, --debug or
--verbose) won't help because they will take effect later, when we call
set_auto_log_levels.

This patch moves the call to set_auto_log_levels earlier, as soon as we
know what the default log level will be.  For all commands but
`convert`, it is after we are done parsing the top-level arguments.  The
default log level will be the most verbose any --debug, --verbose or
--log-level argument passed.

With the `convert` command, it is possible to pass --debug and
--verbose, with the same effect as the top-level equivalents.  So for
this command, we call set_auto_log_levels just after parsing its
arguments.

Change-Id: I9ac558737df4789fc8f5bd8d0240887b3b11b295
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>