Fix: ctf: open metadata file using mode "rb"
In the handling of the support info query, we open the metadata file
using the "r" mode.  On Windows, this is a problem when opening
packetized metadata files, which may contain bytes that Windows will
treat as end of lines and transorm.  Use "rb" instead.

This makes the test_trace_copy test pass on Windows.  It currently fails
with the "session-rotation" trace.

Change-Id: Iaa7a3b8def1c07335c5e7e094d82121df56de268
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>