src/cpp-common: add `bt2_common::TextParseError` class
This patch adds the `bt2_common::TextParseError` class, defined in
`text-parse-error.hpp` and implemented in `text-parse-error.cpp`.

A text parse error is an exception (runtime error) which contains a
stack of text parse error messages. Each text parse error message
contains a string (the message) and a text location to contextualize the
message.

This is part of an effort to implement a JSON parser to support
CTF2‑SPECRC‑4.0 [1]: This new class is needed to implement a string
scanner class and the JSON parser itself.

[1]: https://diamon.org/ctf/files/CTF2-SPECRC-4.0.html

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I056265071dbabf85152d47630817d7a8dfcac7d9
Reviewed-on: https://review.lttng.org/c/babeltrace/+/7409