src/cpp-common: add bt2_common::parseJson() functions (value mode)
This patch adds three new versions of bt2_common::parseJson() which
accept the same parameters as the current "listener mode" versions,
except for a missing listener and an additional, optional `baseOffset`
parameter, and return some resulting JSON value (`bt2_common::JsonVal`).

This is part of an effort to support CTF2‑SPECRC‑5.0 [1]. It will be
easier to convert CTF 2 metadata stream fragments using JSON value
objects than using the listener mode version of bt2_common::parseJson()
directly.

`baseOffset` is added to the offset of a text location when:

1. Creating the text location of any JSON value object to create.

2. Throwing a `TextParseError` instance (from the listener mode
   parseJson() function).

`baseOffset` will be needed when parsing a CTF 2 fragment: the line and
column numbers are local to the fragment, but the offset is from the
beginning of the whole metadata stream. Kind of confusing, but I
wouldn't know what an RS byte means in terms of lines/columns anyway
(different text editor could interpret it in different ways).

[1]: https://diamon.org/ctf/files/CTF2-SPECRC-5.0.html

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I3e9662f6632d10f11c08c178342785886f664797
Reviewed-on: https://review.lttng.org/c/babeltrace/+/7480
Reviewed-by: Simon Marchi <simon.marchi@efficios.com>