configure: enable -Wshadow-field-in-constructor diagnostic
GCC's -Wshadow warns about contructor parameters with the same name as
fields, whereas Clang's does not.  This is a bit annoying when working
with Clang, and then testing with GCC.  Since we'll want to fix these
warnings anyway, so that the code builds with GCC, it would be nice to
have those warnings with Clang directly.

Fortunately, there exists a diagnostic for this in Clang,
`-Wshadow-field-in-constructor`.  Add it to the list of diagnostic flags
we try to enable.

Change-Id: Ib91fa2cada85fe36ec04c6139c8dae1608cf9ed9
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/8024
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>