ctf: use LogCfg throughout `src.ctf.fs`, `src.ctf.lttng-live` and common
A lot of functions and structures have parameters or fields of type
bt_self_component, bt_self_component_class and bt_logging_level for the
sole purpose of logging.  This becomes quite verbose, especially when a
function can be used in both self component and self component class
context, then it will have both of these parameters.

The `ctf::LogCfg` class is used to encapsulate this context in a single
value that can be passed around.  Use that class as much as practically
possible throughout the sources of the ctf plugin, as well as in the
common directory that they depend on.

Remove meta_log_config, which was a structure serving the same purpose.

Change-Id: I0f6844d2cb83d2bf52b81ae9dee4179c7b1f2e12
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>