fixup! lib: add field location-related functions
This was using bt_object_put_ref_no_null_check for no good reason.  It
breaks when destroying a dynamic array field class that has not been
added to a scope, and therefore has not had its path field resolved
(length_field.path.path is NULL).  And in general, destroy functions
should be able to deal with incomplete objects, in case an error happens
mid-construction.

Change-Id: Ie2b3166e9977c20e1318e2b0f2730123d89cbc64