src.ctf.lttng-live: lttng_live_metadata_update: use std::vector as metadata buffer
Use an std::vector to store the metadata contents in
lttng_live_metadata_update, instead of opening a memstream.  Modify
lttng_live_get_one_metadata_packet to append to a vector instead of
writing to a `FILE *`.  We still open a read-only memstream around the
vector's contents to pass a `FILE *` to
ctf_metadata_decoder_append_content.  Change that variable to be a
FileUP, so it's automatically closed on exit.

Change-Id: Ie7b57d25ea90fc82fb96bf19ff707edccdf46072
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/8412
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>